requirements.txt
boto3==1.21.23
celery==5.1.2
coverage==5.5
django-debug-toolbar==3.2.1
django-filter==21.1
django-notifications-hq==1.7.0
django-ratelimit==3.0.1
django-storages==1.12.3
Django==3.1.3
djangorestframework==3.12.4
mysqlclient==2.0.3
pylint==2.11.1
python-memcached==1.59
redis==3.5.3
